# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HTCMC PCHSystem",
    "version": "0.10.0"
  },
  "paths": {
...
    "/sheets/{sheet_id}/managers": {
      "get": {
        "tags": [
          "sheets"
        ],
        "summary": "协管员列表",
        "description": "透明只读：任意登录账号可看（player-less 托管账号亦可，issue #74）。",
        "operationId": "list_managers_sheets__sheet_id__managers_get",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SheetManagerEntry"
                  },
                  "title": "Response List Managers Sheets  Sheet Id  Managers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sheets"
        ],
        "summary": "授予协管员",
        "description": "授予协管员（account 锚）。\n\n- 非 tier A → 403；archived → 409。\n- target Player 不存在或 ``web_account_id is None`` → 422（B7）。\n- target account == owner account → 409 ``SheetOwnerCannotBeManager``（B7）。\n- 幂等重授（is_new=False）不重发通知（PK 防重复）。",
        "operationId": "grant_manager_sheets__sheet_id__managers_post",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerGrantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SheetManagerEntry"
                  },
                  "title": "Response Grant Manager Sheets  Sheet Id  Managers Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sheets"
        ],
        "summary": "移除协管员",
        "description": "撤销协管员（account 锚）。\n\n- self-revoke 放行当 ``player.web_account_id is not None`` 且等于\n  ``body.web_account_id``（B6：显式拒 ``None==None`` 误匹配）。\n- 否则需 tier A；archived → 409；manager 不存在 → 404。\n- self-revoke 不通知（发起方即受影响方）；其他撤销通知 target 账号全部 UUID。",
        "operationId": "revoke_manager_sheets__sheet_id__managers_delete",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerRevokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SheetManagerEntry"
                  },
                  "title": "Response Revoke Manager Sheets  Sheet Id  Managers Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManagerGrantRequest": {
        "properties": {
          "player_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Player Uuid"
          }
        },
        "type": "object",
        "required": [
          "player_uuid"
        ],
        "title": "ManagerGrantRequest",
        "description": "``POST /sheets/{id}/managers``：授予协管员。\n\n按 player_uuid 授予（MCDR 端先用 uuid_api_remake 把玩家名转 UUID 再调本端点）。\n后端解析 target Player → ``web_account_id``；未绑 Web 账号 → 422（B7）。\n目标 account == owner account → 409 ``SheetOwnerCannotBeManager``（B7）。"
      },
      "ManagerRevokeRequest": {
        "properties": {
          "web_account_id": {
            "type": "integer",
            "title": "Web Account Id"
          }
        },
        "type": "object",
        "required": [
          "web_account_id"
        ],
        "title": "ManagerRevokeRequest",
        "description": "``DELETE /sheets/{id}/managers``：撤销协管员（account 锚）。\n\n收 ``web_account_id``（非 player_uuid）以匹配存储模型；同一账号下任一 UUID\n都可 self-revoke（B6 守卫：``player.web_account_id is not None`` 显式拒\nNone==None 误匹配——未绑账号玩家不能 self-revoke）。"
      },
...
      "SheetManagerEntry": {
        "properties": {
          "web_account_id": {
            "type": "integer",
            "title": "Web Account Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "member_uuids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Member Uuids"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          }
        },
        "type": "object",
        "required": [
          "web_account_id",
          "display_name",
          "member_uuids",
          "granted_at"
        ],
        "title": "SheetManagerEntry",
        "description": "项目级协管员（迁移 0014，account 锚，R-5 落地）。\n\n- ``web_account_id``：manager 锚定的 Web 账号 id（PK 一员）。\n- ``display_name``：账号显示名（``WebAccount.display_name`` 优先，否则账号下\n  ``last_seen_at`` 最新 UUID 的 ``current_name``）——经\n  ``web_account_repo.resolve_account_briefs`` 解析。\n- ``member_uuids``：账号下全部 UUID（含 inactive）；客户端按 viewer_uuids 做\n  交集判定 ``is_manager``（前端 auth store 持有绑定 UUIDs；MCDR sheet detail\n  持有 viewer_uuids），无需感知 account_id。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "X-Service-Token": {
        "type": "apiKey",
        "description": "服务端组件令牌（MCDR / 服务端 mod / 服主脚本）",
        "in": "header",
        "name": "X-Service-Token"
      },
      "Authorization": {
        "type": "apiKey",
        "description": "Web JWT 通道，值形如 \"Bearer <jwt>\"",
        "in": "header",
        "name": "Authorization"
      },
      "X-Player-UUID": {
        "type": "apiKey",
        "description": "MCDR 代理通道玩家 UUID（与 X-Service-Token 成对）",
        "in": "header",
        "name": "X-Player-UUID"
      },
      "X-Source-Id": {
        "type": "apiKey",
        "description": "施工上报源标识（服务端 mod 白名单名）",
        "in": "header",
        "name": "X-Source-Id"
      }
    }
  }
}